Created components folder
