+++ packages/strapi-helper-plugin/lib/src/components/Wysiwyg/components/index.js
@@ -1,0 +1,21 @@
+/**
+ *
+ * Wysiwyg components' index
+ *
+ */
+
+import CustomSelect from './customSelect';
+import Image from './image';
+import Link from './link';
+import PreviewControl from './previewControl';
+import PreviewWysiwyg from './previewWysiwyg';
+import ToggleMode from './toggleMode';
+
+export {
+  CustomSelect,
+  Image,
+  Link,
+  PreviewControl,
+  PreviewWysiwyg,
+  ToggleMode,
+};
