+++ packages/strapi-helper-plugin/lib/src/components/Wysiwyg/components/link.js
@@ -1,0 +1,30 @@
+/**
+ *
+ * Link
+ *
+ */
+
+import React from 'react';
+import PropTypes from 'prop-types';
+
+const Link = props => {
+  const { url } = props.contentState.getEntity(props.entityKey).getData();
+
+  return (
+    <a href={url}>
+      {props.children}
+    </a>
+  );
+};
+
+Link.defaultProps = {
+  children: '',
+};
+
+Link.propTypes = {
+  children: PropTypes.node,
+  contentState: PropTypes.object.isRequired,
+  entityKey: PropTypes.string.isRequired,
+};
+
+export default Link;
