+++ packages/strapi-helper-plugin/lib/src/components/Wysiwyg/components/styles.scss
@@ -1,0 +1,117 @@
+.editor {
+  min-height: 303px;
+  padding: 10px 20px 0 20px;
+  font-size: 16px;
+  margin-top: 10px;
+  background-color: #fff;
+  line-height: 18px !important;
+  cursor: text;
+  // TODO define rules for header's margin
+  h1, h2, h3, h4, h5, h6 {
+    margin: 0;
+    line-height: 18px !important;
+  }
+  h1 {
+    margin-top: 13px !important;
+    margin-bottom: 22px;
+  }
+  > div {
+    > div {
+      > div {
+        margin-bottom: 32px;
+      }
+    }
+  }
+  li {
+    margin-top: 0;
+  }
+  ul, ol {
+    margin-bottom: 18px;
+  }
+  span {
+    white-space: pre-line;
+  }
+}
+
+.editorFullScreen {
+  max-height: calc(100% - 100px) !important;
+  margin-bottom: 0;
+  overflow: auto;
+}
+
+.editorInput {
+  height: 0;
+  width: 0;
+}
+
+.editorSelect {
+  min-width: 161px;
+  margin-left: 15px;
+  margin-right: 5px;
+  > select {
+    box-shadow: 0 0 0 rgba(0,0,0,0)!important;
+  }
+}
+
+.fullscreenPreviewEditor {
+  margin-top: 9px;
+  padding: 10px 20px;
+}
+
+.previewControlsWrapper {
+  display: flex;
+  height: 49px;
+  width: 100%;
+  padding: 0 17px;
+  justify-content: space-between;
+  background-color: #FAFAFB;
+  line-height: 30px;
+  font-size: 12px;
+  font-family: Lato;
+  background-color: #fff;
+  border-bottom: 1px solid #F3F4F4;
+  line-height: 49px;
+  font-size: 13px;
+  >div:first-child {
+    >span:last-child {
+      font-size: 12px;
+    }
+  }
+}
+
+.selectFullscreen {
+  min-width: 115px;
+  margin-left: 15px;
+  > select {
+    min-width: 110px !important;
+    box-shadow: 0 0 0 rgba(0,0,0,0)!important;
+  }
+}
+
+.toggleModeButton {
+  height: 32px;
+  min-width: 32px;
+  padding-left: 10px;
+  padding-right: 10px;
+  border: 1px solid rgba(16,22,34,0.10);
+  border-radius: 3px;
+  background: #F3F4F4;
+  font-size: 13px;
+  font-weight: 500;
+  text-align: center;
+  cursor: pointer;
+}
+
+.toggleModeWrapper {
+  margin-left: auto;
+  margin-right: 15px;
+  padding-top: 8px;
+}
+
+.wysiwygCollapse {
+  &:after {
+    content: '\f066';
+    font-family: FontAwesome;
+    margin-left: 8px;
+  }
+}
